Fix review feedback
