+++ dhis-2/dhis-services/dhis-service-validation/src/main/java/org/hisp/dhis/outlierdetection/util/OutlierDetectionUtils.java
@@ -102,4 +102,22 @@ public class OutlierDetectionUtils {
       throw new QueryRuntimeException(E7131);
     }
   }
+
+  /**
+   * Regex pattern to identify strings that are valid for casting to PostgreSQL `double precision`
+   * (i.e., plain numeric values).
+   *
+   * <p>Pattern: ^-?[0-9]+(\.[0-9]+)?$
+   *
+   * <p>Matches: - "42" - "-3.14" - "0.5" - "0001.00"
+   *
+   * <p>Does not match: - "+5" - ".5" - "42." - "1e5" - " 5.5 " - "1,000.00"
+   *
+   * <p>This pattern is used to pre-filter text-based numeric values to avoid runtime casting
+   * exceptions when converting to double precision. Since data values are stored as strings in the
+   * database, there is no guarantee that the string representation of a number is valid for casting
+   * to double precision. Some edge cases may not be covered by this pattern, but integrity checks
+   * can help to identify such cases and fix them.
+   */
+  public static final String PG_DOUBLE_REGEX = "^[+-]?([0-9]*[.])?[0-9]+$";
 }
