+++ dhis-2/dhis-services/dhis-service-validation/src/test/java/org/hisp/dhis/outlierdetection/util/OutlierDetectionUtilsTest.java
@@ -34,10 +34,13 @@ import static org.hisp.dhis.feedback.ErrorCode.E7131;
 import static org.hisp.dhis.outlierdetection.util.OutlierDetectionUtils.withExceptionHandling;
 import static org.hisp.dhis.test.TestBase.createOrganisationUnit;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.google.common.collect.Lists;
 import java.util.List;
+import java.util.regex.Pattern;
 import org.hisp.dhis.common.IllegalQueryException;
 import org.hisp.dhis.common.QueryRuntimeException;
 import org.hisp.dhis.feedback.ErrorCode;
@@ -87,5 +90,36 @@ class OutlierDetectionUtilsTest {
     }
 
     return null;
+  }
+
+  private static final Pattern NUMERIC_PATTERN =
+      Pattern.compile(OutlierDetectionUtils.PG_DOUBLE_REGEX);
+
+  @Test
+  void shouldMatchValidNumbers() {
+    assertTrue(matches("42"));
+    assertTrue(matches("0"));
+    assertTrue(matches("-123"));
+    assertTrue(matches("3.14"));
+    assertTrue(matches("0.0"));
+    assertTrue(matches("0001.00"));
+    assertTrue(matches("-0.5"));
+  }
+
+  @Test
+  void shouldNotMatchInvalidNumbers() {
+    assertFalse(matches("+42"));
+    assertFalse(matches(".5"));
+    assertFalse(matches("42."));
+    assertFalse(matches("1e5"));
+    assertFalse(matches("1,000"));
+    assertFalse(matches("abc"));
+    assertFalse(matches(""));
+    assertFalse(matches(null));
+  }
+
+  private boolean matches(String value) {
+    if (value == null) return false;
+    return NUMERIC_PATTERN.matcher(value.trim()).matches();
   }
 }
